// src/api/types/schema/models/ApiCommitRowDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiCheckpointDto } from './ApiCheckpointDto';
import type { ApiCommitDto } from './ApiCommitDto';
export type ApiCommitRowDto = {
    checkpoint: ApiCheckpointDto;
    commit: ApiCommitDto;
};